Allow HDF5 to be compiled statically.
The developers behind the HDF5 family of crates have a nice feature that
should be a little more visible - it's possible to have the HDF5 library
compiled automatically by cargo, such that it is not needed as a system
library. This also allows python wheels to be built for distribution(!).
Add automated testing and compilation for GitHub actions.

I don't think it's possible to automatically release the compiled Python
wheels from GitHub to PyPI. For now, this should be done manually (i.e.
download the wheel from GitHub, upload with twine).

Shuffled and tidied up some code. The main reasons for moving code
around are that analytic beam code should be in hyperbeam eventually,
and I spotted a couple of things that lingered from the initial commit.

Allow docs.rs to build.