Rust API changes.
The parallactic angle correction no longer has a hard-coded latitude. If
the caller wants the correction, the latitude must be provided,
otherwise it is not done. Along with this, remove "_eng" functions from
CPU code (their results are still available if the array latitude is not
given).

Add a new flag `iau_order` to control the polarisation order of the
beam-response Jones matrices after parallactic-angle correction. The
details are written up in a repo-tracked fee_pols.pdf.

When taking in delays as `&[u32]`, throw an error if this doesn't have a
length of 16. When taking in delays as `ArrayView2<u32>`, throw a
similar error if this doesn't have 16 columns. Also add a similar error
for the number of amps provided. Remove all old debug asserts associated
with delay lengths.

Reduce the number of unwraps when getting delays from ndarrays. Also
remove an unneeded unwrap in the init function.

Move Rust examples into a new crate.