Update dependencies.
We no longer need the ERFA C library.

A bunch of changes have been made due to the new Marlu. Explicitly depending on
ndarray and rayon isn't necessary (can get through through Marlu), but I prefer
this to be explicit.

Turn off default features for cbindgen; this stops clap 3 being pulled in as a
hard dependency.