CPU code improvements.
When working on the CUDA code in the last commit, I realised that the
"p1 arrays" can be shared between X and Y dipoles. Using that info made
the CUDA code significantly faster. This commit does the same thing for
the CPU code, and it also now runs significantly faster.

Depend on num-complex 0.4.1 explicitly; this allows us to use
Complex::cis, which is much tidier than the manual implementation used
in the code.

Re-use some vectors rather than re-creating them for every iteration in
a couple of loops. It's possible that the allocator was smart enough to
not do too much extra work but this makes it explicit. Also shrink the
capacity of coefficient vectors before storing them in the cache.

Optimise `FEEBeam::find_closest_freq`; a benchmark shows this went from 99ns to
75ns.

Change the type of n_max to u8 from usize.

Use u32::abs_diff instead of the manual equivalent.

Tidy up use statements.