Set CXX for CUDA, if conditions are met.
Thanks to Dev for the suggestion.

We've had a long-standing "issue" with Archlinux that, because it distributes
really new compilers, these may not play nicely with what nvcc expects. To get
around that, there's a gcc and g++ in $CUDA_PATH/bin. I've manually gotten
around compilation issues by setting $CXX to this, but now $CXX will be set if:

1) $CXX is not already set
2) $CUDA_PATH is set
3) $CUDA_PATH/bin/g++ exists