Fix a nasty race condition.
This commit reflects Dev's discovery of the problem -- thanks Dev.

In these few instances of `Option<DevicePointer<GpuFloat>>`, the pointer within
a `DevicePointer` was being taken, but then the `DevicePointer` itself was
consumed, so if the relevant memory was altered, the code did not behave as
intended. Curiously, this never led to a segfault despite being a use after
free, and witnessing unintended behaviour seemed to require running many GPU
tests concurrently.

The code is now simpler and less prone to error; don't have an
`Option<DevicePointer>`, just make a `DevicePointer` and if it has no contents,
it will give out a null pointer anyway.